requirements.txt
streamlit
transformers
pdf2docx
lxml
torch
python-dotenv
tempfile
shutil
logging